+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -692,6 +692,7 @@ class ConfigurationTest extends TestCase
                 'enabled' => true,
                 'enable_attributes' => !class_exists(FullStack::class),
                 'mapping' => ['paths' => []],
+                'named_serializers' => [],
             ],
             'property_access' => [
                 'enabled' => true,
@@ -866,5 +867,22 @@ class ConfigurationTest extends TestCase
                 'enabled' => false,
             ],
         ];
+    }
+
+    public function testNamedSerializersReservedName()
+    {
+        $processor = new Processor();
+        $configuration = new Configuration(true);
+
+        $this->expectException(InvalidConfigurationException::class);
+        $this->expectExceptionMessage('Invalid configuration for path "framework.serializer.named_serializers": "default" is a reserved name.');
+
+        $processor->processConfiguration($configuration, [[
+            'serializer' => [
+                'named_serializers' => [
+                    'default' => ['include_standard_normalizers' => false],
+                ],
+            ],
+        ]]);
     }
 }